moved algorithm stuff to hash package and fixed typo in README
