+++ src/main/java/gov/loc/repository/bagit/domain/Manifest.java
@@ -4,6 +4,8 @@ package gov.loc.repository.bagit.domain;
 import java.io.File;
 import java.util.HashMap;
 
+import gov.loc.repository.bagit.hash.SupportedAlgorithm;
+
 /**
  * A manifest is a list of files and their corresponding checksum with the algorithm used to generate that checksum
  */
