+++ src/test/java/gov/loc/repository/bagit/writer/BagWriterTest.java
@@ -18,8 +18,8 @@ import org.junit.rules.TemporaryFolder;
 import gov.loc.repository.bagit.domain.Bag;
 import gov.loc.repository.bagit.domain.FetchItem;
 import gov.loc.repository.bagit.domain.Manifest;
-import gov.loc.repository.bagit.domain.StandardSupportedAlgorithms;
 import gov.loc.repository.bagit.domain.Version;
+import gov.loc.repository.bagit.hash.StandardSupportedAlgorithms;
 import gov.loc.repository.bagit.reader.BagReader;
 
 public class BagWriterTest extends Assert {
